# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentHub REST API
  version: v1
  description: >
    AgentHub REST JSON API for commands and queries. Realtime state, logs,
    message deltas, approvals and artifact notifications are delivered through
    WebSocket typed events documented in api/events.md.

    ## Implementation Status (2026-05-29)

    **Edge Server**: Core endpoints (health/runs/threads/projects/agents/events/permissions)
    are implemented. `/v1/runners` remains a compatibility surface for runtime/
    target health; new architecture language should use Agent Runtime, Agent
    Profile, and Execution Target.

    **Hub Server**: Hub entries marked `x-agenthub-status: implemented` mirror
    the current Gin router paths in `hub-server/internal/router/router.go`,
    including `/client/auth`, `/client/contacts`, `/client/sessions`,
    `/client/messages`, `/client/attachments`, `/client/notifications`,
    `/edge/*`, and `/web/*`. Deprecated or legacy Hub shapes are retained only
    when marked `planned`; they are not callable routes.
    The legacy `/v1/auth`, `/v1/users`, `/v1/contacts` paths are planned for future
    migration and not currently implemented.

    ## Status tags
    - `x-agenthub-status: implemented` — 已实现，可直接调用
    - `x-agenthub-status: planned` — 远期规划，尚未实现
servers:
  - url: http://127.0.0.1:3210
    description: Local Edge Server
  - url: http://localhost:8080
    description: Local Hub Server
  - url: http://api.hub.vectorcontrol.tech
...
paths:
...
  /v1/events:
    get:
      tags: [Foundation]
      operationId: openEventStream
      summary: Open WebSocket event stream.
      description: >
        Local Edge WebSocket handshake surface. Event envelope and event types are
        documented in api/events.md. When local Edge auth is enabled, browser
        clients use `access_token` query auth; native clients may use the same
        local Edge token through Authorization or X-AgentHub-Edge-Token.
      x-agenthub-phase: P0
      x-agenthub-status: implemented
      x-agenthub-owner: Edge
      security:
        - edgeWebSocketQueryToken: []
        - edgeBearerAuth: []
        - edgeLocalToken: []
      parameters:
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "101":
          description: Switching Protocols.
        "401":
          $ref: "#/components/responses/Error"
  /v1/projects:
    get:
      tags: [IMProject]
      operationId: listProjects
      summary: List projects.
      x-agenthub-phase: P0
      x-agenthub-status: implemented
      x-agenthub-owner: Edge
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          $ref: "#/components/responses/List"
    post:
      tags: [IMProject]
      operationId: createProject
      summary: Create or register a project.
      x-agenthub-phase: P0
      x-agenthub-status: implemented
      x-agenthub-owner: Edge
      responses:
        "201":
          $ref: "#/components/responses/Created"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Hub HS256 access token
      description: >
        Hub-issued access token minted by Hub Server after password login or
        TokenDance ID OIDC exchange. Do not send a raw TokenDance ID RS256
        bearer token or a Relay API key to Hub REST/WebSocket endpoints.
    edgeBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Local Edge token
      description: >
        Optional Local Edge token configured by AGENTHUB_EDGE_AUTH_TOKEN or
        --local-auth-token. This is only the loopback Edge process boundary and
        is not a Hub session, TokenDance ID token, or Relay API key.
    edgeLocalToken:
      type: apiKey
      in: header
      name: X-AgentHub-Edge-Token
      description: >
        Header alternative for the optional Local Edge token when Authorization
        is already owned by another caller context.
    edgeWebSocketQueryToken:
      type: apiKey
      in: query
      name: access_token
      description: >
        Browser WebSocket transport for the optional Local Edge token on
        /v1/events. Hub WebSocket uses Hub-issued access tokens separately.
    hubWebSocketQueryToken:
      type: apiKey
      in: query
      name: access_token
      description: >
        Hub-issued HS256 access token used by browser clients during /client/ws
        HTTP upgrade. TokenDance ID bearer tokens and TokenDance API keys are
        not accepted.
  parameters:
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    PageCursor:
      name: pageCursor
      in: query
      required: false
      schema:
        type: string
...
  schemas:
    Resource:
      type: object
      additionalProperties: true
    ListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        page:
          $ref: "#/components/schemas/PageInfo"
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor:
          type: string
        hasMore:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            traceId:
              type: string
            details:
              type: object
              additionalProperties: true
...
  responses:
...
    Created:
      description: Created.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Resource"
...
    List:
      description: List response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ListResponse"
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"